Header section arranged and add new properties
Alphabetize the order on properties for organization of properties.

Change design flow to the overall header section to be the same level on both header and main title.

Change some section to wrap in the container to apply some dynamic flow when the zoom level is changed.

Capitalize the hex color for consistency of the values.